sql/planner/binder/mod.rs -> binder.rs
